refactor(hidpp): split channel.rs and dedup the descriptor probe
`channel.rs` was 1422 lines holding four separable concerns: the transport
trait and its HID++ probe, the raw message framing, the error type, and the
channel itself. It is now a module root at 563 lines with `channel/raw.rs`,
`channel/message.rs`, `channel/error.rs`, and `channel/tests.rs` beneath it.

The probe carried the same 16-line descriptor walk twice, once per report
width, differing only in the report id and usage. It is one
`declares_report` helper called twice. The four usage constants it needs
(`SHORT_REPORT_USAGE_PAGE` and friends) had no user inside the workspace or
out; they are now private to `raw.rs`, with the page — identical for both
widths — stated once instead of twice.

`from_raw_channel` no longer carries the read loop inline. Sixty lines
nested four deep inside a `thread::spawn` closure are now `read_loop`, a
named function whose parameters state exactly what the thread shares with
the channel.

`short_payload_as_long` becomes `HidppMessage::widened`, so widening is
expressed on the message rather than on a loose payload array, and
`normalize_outgoing` stops destructuring a payload only to rebuild it. No
behaviour change; its test now goes through the message, and a second test
pins that widening an already-long message is a no-op.
